requirements.txt
annotated-types==0.7.0
anthropic==0.36.0
anyio==4.6.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.9
distro==1.9.0
Django==5.1.2
django-appconf==1.0.6
django-compressor==4.5.1
django-environ==0.11.2
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.9.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-documentai==2.33.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
iniconfig==2.0.0
jiter==0.6.1
Levenshtein==0.26.1
mimetypes-magic==0.4.30
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
